pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools",
]

[project]
name = "prometheus-aioexporter"
description = "Asyncio library for creating Prometheus exporters"
readme = "README.rst"
keywords = [
  "exporter",
  "library",
  "prometheus",
]
license = { file = "LICENSE.txt" }
maintainers = [
  { name = "Alberto Donato", email = "alberto.donato@gmail.com" },
]
authors = [
  { name = "Alberto Donato", email = "alberto.donato@gmail.com" },
]
requires-python = ">=3.11"
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Utilities",
]
dynamic = [
  "version",
]
dependencies = [
  "aiohttp>=3",
  "click>=8.2",
  "prometheus-client>=0.4",
  "python-dotenv",
  "structlog",
]
optional-dependencies.testing = [
  "pytest",
  "pytest-aiohttp",
  "pytest-asyncio",
  "pytest-mock",
  "pytest-structlog",
  "trustme",
]
urls.changelog = "https://github.com/albertodonato/prometheus-aioexporter/blob/main/CHANGES.rst"
urls.homepage = "https://github.com/albertodonato/prometheus-aioexporter"
urls.repository = "https://github.com/albertodonato/prometheus-aioexporter"
scripts.prometheus-aioexporter-sample = "prometheus_aioexporter.sample:script"

[tool.setuptools.dynamic]
version = { attr = "prometheus_aioexporter.__version__" }

[tool.setuptools.packages.find]
include = [ "prometheus_aioexporter*" ]

[tool.setuptools.package-data]
prometheus_aioexporter = [ "py.typed" ]

[tool.ruff]
line-length = 79

lint.select = [ "F", "I", "RUF", "UP" ]
lint.isort.combine-as-imports = true
lint.isort.force-sort-within-sections = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
structlog_keep = [ "TimeStamper", "add_log_level", "JSONRenderer" ]

[tool.coverage.report]
fail_under = 100.0
show_missing = true
skip_covered = true

[tool.coverage.run]
source = [ "prometheus_aioexporter" ]
omit = [ "prometheus_aioexporter/sample.py" ]

[tool.mypy]
install_types = true
non_interactive = true
strict = true
